backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
